// app/interfaces/ClientInfographics.tsx
  export interface ClientInfographic {
    id: number;
    payer_id: string;
    client_name: string;
    title: string;
    email: string;
    address: string;
    fullDescription: string;
    created: string;
    facebook: string;
    facebookUrl: string;
    twitter: string;
    twitterUrl: string;
    instagram: string;
    instagramUrl: string;
    paypalTransaction: string;
    amount: number;
    currency: string;
    state: string;
    type: string;
    imgUrl: string;
    liveUrl: string;
    publishDate: string;
    paypalId: string;
    status: number;
    catId: number;
  }